Removed : - Disassembler code - libretro's specific code. (Flash code will need some specific changes though later on) - Switched most of the code to C except for the Audio code. - Save states code is now completely replaced too.
Still need to work on the menu and separating some of the code
in libretro.c but we're almost there.

+++ mednafen/ngp/rom.c
@@ -11,11 +11,10 @@
 //	(at your option) any later version. See also the license.txt file for
 //	additional informations.
 //---------------------------------------------------------------------------
-
+#include <stdint.h>
 #include "neopop.h"
 #include "flash.h"
 #include "interrupt.h"
-#include "TLCS-900h/TLCS900h_disassemble.h"
 #include "../mednafen.h"
 
 #ifdef MSB_FIRST
@@ -105,7 +104,7 @@ void rom_unload(void)
 {
    if (ngpc_rom.data)
    {
-      int i;
+      uint_fast8_t i;
 
       flash_commit();
